services/normalizer-gtfsrt/requirements.txt
kafka-python==2.0.2
protobuf>=5.27.0,<6
python-dotenv==1.0.1
gtfs-realtime-bindings
